+++ src/_types/dehydrated_val__T.d.ts
@@ -1,6 +1,5 @@
 import type { data_row_T } from './data_row_T'
 import type { dehydrated_json_val_T } from './dehydrated_json_val_T'
-import type { hydrated_json_val_T } from './hydrated_json_val_T'
 export type dehydrated_val__T<
 	ColDefs extends (([string, any][])|any[]|object) = ([string, any][])|any[]|object,
 	D = dehydrated_json_val_T