+++ src/data_row__dehydrate/index.js
@@ -8,18 +8,18 @@ import { json_val__dehydrate } from '../json_val__dehydrate/index.js'
 /** @typedef {import('../_types').hydrated_json_val_T}hydrated_json_val_T */
 /**
  * @param {data_row_T}data_row
- * @param {dehydrated_val__T}dehydrated_val_
+ * @param {dehydrated_val__T}val__dehydrate
  * @returns {dehydrated_json_data_row_tuple_T}
  * @private
  */
 export function data_row__dehydrate(
 	data_row,
-	dehydrated_val_ = json_val__dehydrate
+	val__dehydrate = json_val__dehydrate
 ) {
 	/** @type {header_row_T} */
 	const header_row = data_row[header_M_col_idx_sym].header_row
 	return data_row.map((hydrated_json_val, col_idx)=>
-		dehydrated_val_(
+		val__dehydrate(
 			hydrated_json_val,
 			header_row[col_idx],
 			data_row,